+++ src/config.php
@@ -2,21 +2,19 @@
 /**
  * Minify plugin for Craft CMS 3.x
  *
- * Some Description
- *
- * @link      https://nystudio107.com
+ * @link      https://nystudio107.com/
  * @copyright Copyright (c) 2017 nystudio107
+ * @license   MIT License https://opensource.org/licenses/MIT
  */
 
 /**
  * Minify config.php
  *
- * Completely optional configuration settings for Minify if you want to
- * customize some of its more esoteric behavior, or just want specific control
- * over things.
+ * This file exists only as a template for the Minify settings.
+ * It does nothing on its own.
  *
- * Don't edit this file, instead copy it to 'craft/config' as 'minify.php' and
- * make your changes there.
+ * Don't edit this file, instead copy it to 'craft/config' as 'minify.php'
+ * and make your changes there to override default settings.
  *
  * Once copied to 'craft/config', this file will be multi-environment aware as
  * well, so you can have different settings groups for each environment, just as
@@ -29,6 +27,6 @@ return [
     "disableTemplateMinifying" => false,
 
     // if set to `true` then Minify will not minify anything if `devMode` is enabled
-    "disableDevmodeMinifying" => false,
+    "disableDevModeMinifying" => false,
 
 ];