+++ src/models/Settings.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * Minify plugin for Craft CMS 3.x
+ *
+ * @link      https://nystudio107.com/
+ * @copyright Copyright (c) 2017 nystudio107
+ * @license   MIT License https://opensource.org/licenses/MIT
+ */
+
+namespace nystudio107\minify\models;
+
+use nystudio107\minify\Minify;
+
+use Craft;
+use craft\base\Model;
+
+/**
+ * Minify Settings model
+ *
+ * @author    nystudio107
+ * @package   Minify
+ * @since     1.2.0
+ */
+class Settings extends Model
+{
+    // Public Properties
+    // =========================================================================
+
+    /**
+     * If set to `true` then Minify will not minify anything
+     *
+     * @var bool
+     */
+    public $disableTemplateMinifying = false;
+
+    /**
+     * If set to `true` then Minify will not minify anything if `devMode` is enabled
+     *
+     * @var bool
+     */
+    public $disableDevModeMinifying = false;
+
+
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    public function rules()
+    {
+        return [
+            ['disableTemplateMinifying', 'boolean'],
+            ['disableTemplateMinifying', 'default', 'value' => false],
+            ['disableDevModeMinifying', 'boolean'],
+            ['disableDevModeMinifying', 'default', 'value' => false],
+        ];
+    }
+}